Reinstate use of has_outcome_per_path function
This function was created in an earlier pull request to manage the list of
countries that use an outcome template per path through the flow.

The use of the function must have been lost during a rebase, so I'm
reinstating it here.

No test artefacts were affected by this change.